\name{irisCourse-package}
\alias{irisCourse-package}
\alias{irisCourse}
\docType{package}
\title{
\packageTitle{irisCourse}
}
\description{
\packageDescription{irisCourse}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{irisCourse}
\packageIndices{irisCourse}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{irisCourse}

Maintainer: \packageMaintainer{irisCourse}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
